maybe workign can.
can FD packets seem to get mostly through but everything over 8B fails. THinks its bcause one of the transceivers im using is not rated for CAN-FD. though it does the same even when there is not bit-rate switching... idk what in the chip is different for one that supports CAN-FD or not. must be somethign cool